requirements.txt
aiogram==3.4.1
aiohttp>=3.9.0
SQLAlchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
openai==1.12.0
APScheduler==3.10.4
loguru==0.7.2
python-dotenv==1.0.1
pydantic<2.6
pydantic-settings==2.1.0
